refactor: replace useContactEmailDuplicateCheck with inline logic for email existence check
